@@ -67,95 +74,29 @@
     "autoprefixer": "^10.4.20",
     "babel-loader": "^9.2.1",
     "babel-preset-react-app": "^10.0.1",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.19.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.18",
     "eslint-plugin-simple-import-sort": "^12.1.1",
+    "eslint-plugin-storybook": "^0.11.2",
+    "globals": "^15.14.0",
     "happy-dom": "^16.7.3",
     "postcss": "^8.5.1",
-    "pre-commit": "^1.2.2",
     "prettier": "^3.4.2",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "prettier-plugin-packagejson": "^2.5.8",
+    "prettier-plugin-tailwindcss": "^0.6.11",
     "prop-types": "^15.8.1",
-    "storybook": "^8.4.7",
+    "storybook": "^8.5.2",
     "tailwindcss": "^3.4.17",
+    "typescript-eslint": "^8.22.0",
     "vite": "^6.0.11",
     "vite-plugin-eslint": "^1.8.1",
-    "vitest": "^2.1.8"
-  },
-  "pre-commit": "fix",
-  "browserslist": {
-    "production": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "prettier": {
-    "plugins": [
-      "prettier-plugin-organize-imports",
-      "prettier-plugin-packagejson",
-      "prettier-plugin-tailwindcss"
-    ],
-    "semi": true,
-    "singleQuote": true,
-    "printWidth": 120,
-    "tailwindFunctions": [
-      "clsx"
-    ]
-  },
-  "eslintConfig": {
-    "env": {
-      "browser": true,
-      "es2021": true,
-      "jest": true,
-      "node": true
-    },
-    "parserOptions": {
-      "ecmaFeatures": {
-        "jsx": true
-      },
-      "ecmaVersion": 2022,
-      "sourceType": "module"
-    },
-    "settings": {
-      "react": {
-        "version": "detect"
-      }
-    },
-    "plugins": [
-      "react",
-      "jsx-a11y",
-      "react-hooks",
-      "react-refresh",
-      "prettier"
-    ],
-    "extends": [
-      "eslint:recommended",
-      "plugin:react/recommended",
-      "plugin:react/jsx-runtime",
-      "plugin:react-hooks/recommended",
-      "plugin:jsx-a11y/recommended",
-      "plugin:storybook/recommended",
-      "plugin:prettier/recommended"
-    ],
-    "rules": {
-      "prettier/prettier": [
-        "error"
-      ],
-      "react-refresh/only-export-components": [
-        "warn"
-      ],
-      "react/prop-types": [
-        "off"
-      ]
-    }
+    "vitest": "^3.0.4"
   }
 }
